ui components: add shimmer loading effect to components
