+++ Modules/feature-common/Sources/UI/Success/Document/DocumentSuccessView.swift
@@ -92,9 +92,9 @@ private func documents<RequestItem: Sendable>(
           items: section.listItems,
           backgroundColor: backgroundColor,
           hideSensitiveContent: false,
+          isLoading: viewState.isLoading,
           onItemClick: { onSelectionChanged($0.groupId) }
         )
-        .shimmer(isLoading: viewState.isLoading)
       }
     }
   }
