+++ Modules/feature-dashboard/Sources/UI/Detail/Transaction/TransactionDetailsView.swift
@@ -60,7 +60,8 @@ private func content(
 
       if let transactionDetailsCardData = state.transactionDetailsUi?.transactionDetailsCardData {
         TransactionCardView(
-          transactionDetailsCardData: transactionDetailsCardData
+          transactionDetailsCardData: transactionDetailsCardData,
+          isLoading: state.isLoading
         )
       }
 
@@ -70,6 +71,7 @@ private func content(
             .typography(Theme.shared.font.bodySmall)
             .fontWeight(.semibold)
             .foregroundStyle(Theme.shared.color.onSurfaceVariant)
+            .shimmer(isLoading: state.isLoading)
 
           ForEach(transactionDetailsDataSharedList) { item in
             WrapExpandableListView(
@@ -78,7 +80,8 @@ private func content(
                 supportingText: .viewDetails
               ),
               items: item.listItems,
-              hideSensitiveContent: false
+              hideSensitiveContent: false,
+              isLoading: state.isLoading
             )
           }
         }
@@ -88,6 +91,7 @@ private func content(
             Text(.transactionDetailsRequestDeletionMessage)
               .font(Theme.shared.font.bodyLarge.font)
               .foregroundStyle(Theme.shared.color.onSurfaceVariant)
+              .shimmer(isLoading: state.isLoading)
 
             WrapButtonView(
               style: .error,
@@ -102,6 +106,7 @@ private func content(
             Text(.transactionDetailsReportTransactionMessage)
               .font(Theme.shared.font.bodyLarge.font)
               .foregroundStyle(Theme.shared.color.onSurfaceVariant)
+              .shimmer(isLoading: state.isLoading)
 
             WrapButtonView(
               style: .secondary,
@@ -114,7 +119,6 @@ private func content(
         }
       }
     }
-    .shimmer(isLoading: state.isLoading)
     .padding(Theme.shared.dimension.padding)
     .padding(.bottom)
   }
